+++ client/pages/photos.tsx
@@ -8,7 +8,7 @@ import { useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { NextSeo } from 'next-seo'
 import { useRouter } from 'next/router'
-import React, { useMemo, useState } from 'react'
+import React, { useEffect, useMemo, useState } from 'react'
 import { Button, Dropdown, Input } from 'simple-react-ui-kit'
 
 import AppFooter from '@/components/app-footer'
@@ -17,11 +17,13 @@ import AppToolbar from '@/components/app-toolbar'
 import PhotoGrid from '@/components/photo-grid'
 
 interface PhotosPageProps {
+    category: string
     photosList: ApiModel.Photo[]
     categoriesList: ApiModel.Category[]
 }
 
 const PhotosPage: NextPage<PhotosPageProps> = ({
+    category,
     photosList,
     categoriesList
 }) => {
@@ -34,10 +36,6 @@ const PhotosPage: NextPage<PhotosPageProps> = ({
 
     const [searchFilter, setSearchFilter] = useState<string>()
     const [categoryFilter, setCategoryFilter] = useState<number | undefined>()
-
-    const handleCreate = () => {
-        router.push('/photos/form')
-    }
 
     const filteredCategoriesList = useMemo(
         () =>
@@ -72,11 +70,46 @@ const PhotosPage: NextPage<PhotosPageProps> = ({
         [photosList, categoryFilter, searchFilter]
     )
 
+    const currentCategory = useMemo(
+        () => filteredCategoriesList?.find(({ id }) => id === categoryFilter),
+        [filteredCategoriesList, categoryFilter]
+    )
+
+    const title =
+        t('astrophoto') + (categoryFilter ? `: ${currentCategory?.title}` : '')
+
+    const handleChangeCategoryFilter = (category: number | undefined) => {
+        if (category !== undefined) {
+            router.push({
+                pathname: router.pathname,
+                query: { ...router.query, category }
+            })
+        } else {
+            const { ...rest } = router.query
+            router.push({
+                pathname: router.pathname,
+                query: rest
+            })
+        }
+
+        setCategoryFilter(category)
+    }
+
+    const handleCreate = () => {
+        router.push('/photos/form')
+    }
+
+    useEffect(() => {
+        setCategoryFilter(category ? parseInt(category) : undefined)
+    }, [category])
+
     return (
         <AppLayout>
             <NextSeo
-                title={t('astrophoto')}
-                description={t('description-photos-list-page')}
+                title={title}
+                description={`${t('description-photos-list-page')} ${
+                    currentCategory?.description
+                }`}
                 canonical={`${canonicalUrl}photos`}
                 openGraph={{
                     images: [
@@ -92,8 +125,20 @@ const PhotosPage: NextPage<PhotosPageProps> = ({
             />
 
             <AppToolbar
-                title={t('astrophoto')}
-                currentPage={t('astrophoto')}
+                title={title}
+                currentPage={
+                    categoryFilter ? currentCategory?.title : t('astrophoto')
+                }
+                links={
+                    categoryFilter
+                        ? [
+                              {
+                                  link: '/photos',
+                                  text: t('astrophoto')
+                              }
+                          ]
+                        : undefined
+                }
             >
                 <Input
                     placeholder={t('search-by-object')}
@@ -106,7 +151,9 @@ const PhotosPage: NextPage<PhotosPageProps> = ({
                     size={'medium'}
                     value={categoryFilter}
                     placeholder={t('filter-by-category')}
-                    onSelect={(category) => setCategoryFilter(category?.key)}
+                    onSelect={(category) =>
+                        handleChangeCategoryFilter(category?.key)
+                    }
                     options={filteredCategoriesList?.map((category) => ({
                         key: category.id,
                         value: category.title
@@ -135,6 +182,7 @@ export const getServerSideProps = wrapper.getServerSideProps(
     (store) =>
         async (context): Promise<GetServerSidePropsResult<PhotosPageProps>> => {
             const locale = context.locale ?? 'en'
+            const category = (context.query.category as string) || ''
             const translations = await serverSideTranslations(locale)
 
             store.dispatch(setLocale(locale))
@@ -152,6 +200,7 @@ export const getServerSideProps = wrapper.getServerSideProps(
             return {
                 props: {
                     ...translations,
+                    category,
                     photosList: photos?.items || [],
                     categoriesList: categories?.items || []
                 }